🚚🚧 Restructure components and output
