+++ src/components/GlobalComponent.ts
@@ -1,0 +1,10 @@
+import { Component, BaseComponent, Global } from '@jovotech/framework';
+import { LoveHatePizzaComponent } from './LoveHatePizzaComponent';
+
+@Global()
+@Component()
+export class GlobalComponent extends BaseComponent {
+  LAUNCH() {
+    return this.$redirect(LoveHatePizzaComponent);
+  }
+}
